#2 chore : 날짜 포맷 변환용 함수 생성
